% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_weights.R
\name{add_weights}
\alias{add_weights}
\title{Add a column of survey weights to a data frame.
For use in surveys where you took a sample population out of a larger
source population.}
\usage{
add_weights(x, p, cz = NULL, ..., population, cluster_cz, household_cz,
  cluster_x, household_x, individuals_eligible_x,
  individuals_interviewed_x, method = "stratified",
  ignore_cluster = TRUE, ignore_household = TRUE,
  surv_weight = "surv_weight", surv_weight_ID = "surv_weight_ID")
}
\arguments{
\item{x}{a data frame of survey data}

\item{p}{a data frame containing popuplation data for groups in \code{...}}

\item{cz}{a data frame containing a list of clusters and the number of
households in each. Default is NULL as the default \code{method} is "stratified".}

\item{...}{shared grouping columns across both \code{x} and \code{p}. These are used
to match the weights to the correct subset of the pouplation.
Used only when \code{method} is "stratified".}

\item{population}{the column in \code{p} that defines the pouplation numbers}

\item{cluster_cz}{the column in \code{cz} that lists all possible clusters.
Ignored if \code{method} is not "cluster" or if \code{ignore_cluster} is TRUE.}

\item{household_cz}{the column in \code{cz} that lists the number of households per cluster.
Ignored if \code{method} is not "cluster" or if \code{ignore_household} is TRUE.}

\item{cluster_x}{the column in \code{x} that indicates which cluster rows belong to.
Ignored if \code{method} is not "cluster" or if \code{ignore_cluster} is TRUE.}

\item{household_x}{the column in \code{x} that indicates a unique household identifier.
Ignored if \code{method} is not "cluster or if \code{ignore_household} is TRUE.}

\item{individuals_eligible_x}{the column in \code{x} which specifies the number of people
eligible for being interviewed in that household. (e.g. the total number of children)}

\item{individuals_interviewed_x}{the column in \code{x} which specifies the number of people
actually interviewed in that household.}

\item{method}{what type of survey method you would like to use. Options are
"stratified" (default) and "cluster".
\itemize{
\item Stratified: takes variables grouped population counts from \code{p} and divides
the corresponding groups by counts of variables in \code{...} in \code{x}, to create a weight.
\item Cluster: Will multiply the inverse chances of a cluster being selected, a household
being selected within a cluster, and an individual being selected within a household.
As follows:
(unique(cluster_cz) / unique(cluster_x)) *
(unique(household_cz) / unique(household_x)) *
(individuals_eligible_x / individuals_interviewed_x)
In the case where ignore_cluster and ignore_household are TRUE, this will simply be:
1 * 1 * individuals_eligible_x / inidivudals_interviewed_x
}}

\item{ignore_cluster}{If TRUE, set the weight for clusters to be 1. This
assumes that your sample was taken in a way which is a close approximation
of a simple random sample.  Ignores inputs from \code{cluster_cz} as well as
\code{cluster_x}.  Default is TRUE, as the default \code{method} is "stratified".}

\item{ignore_household}{If TRUE, set the weight for households to be 1. This
assumes that your sample of households was takenin a way which is a close
approximation of a simple random sample. Ignores inputs from \code{household_cz}
and \code{household_x}.  Default is TRUE, as the default \code{method} is "stratified".}

\item{surv_weight}{the name of the new column to store the weights. Defaults to
"surv_weight".}

\item{surv_weight_ID}{the name of the new ID column to be created. Defaults to
"surv_weight_ID"}
}
\description{
NOTES TO OURSELVES:
We want to have two options for weighting:
}
\details{
\itemize{
\item Strata: What we already have in previous version.
Just want to weight each person based on the source population.
E.g. if we want to stratify our analysis based on age and sex in two
different camps. Then we need the population breakdown by age and sex
in those two camps. Then the function just creates a multiplier,
by divding the number from the source population by the number in
the sample population.
\item Cluster: This is more complex. Cluster sampling involves first choosing a
number of clusters from a list of all clusters (e.g. pick a few villages
out of all villages in an area),
then within those villages choosing houses (theres several methods for this),
and finally at the household level you either interview one person
or everyone in that household.
So the population no longer matters,
instead of being population counts as in the above stratified example,
you have a list of all clusters in the area with the number of houses
within each cluster.
Your survey dataset is collected with two levels, one for household
and one for individual. But those two levels are merged, so for
each individual there are two variables included: number of people in
household, and number of people surveyed.
To calculate the weight for each individual, you multiply the inverses
(i.e. 1 over) the probability of: a cluster being chosen,
a household being chosen in that cluster, an individual being chosen
in their household.
}
}
\examples{


# define a fake dataset of survey data
# including household and individual information
x <- tibble::tribble(
  ~cluster, ~household_id, ~eligibile_n, ~surveyed_n, ~individual_id, ~age_grp, ~sex, ~outcome,
  "Village A",             1,            6,           4,              1,   "0-10",  "Male",        "Y",
  "Village A",             1,            6,           4,              2,  "20-30",  "Female",      "Y",
  "Village A",             1,            6,           4,              3,  "30-40",  "Male",        "N",
  "Village A",             1,            6,           4,              4,  "50-60",  "Female",      "N",
  "Village A",             2,            6,           4,              4,  "50-60",  "Female",      "N",
  "Village B",             2,            3,           3,              1,  "20-30",  "Male",        "N",
  "Village B",             2,            3,           3,              2,  "50-60",  "Female",      "N",
  "Village B",             2,            3,           3,              3,  "30-40",  "Female",      "Y"
)

# define a fake population data set
# including age group, sex, counts and proportions
p <- sitrep::gen_population(total = 10000,
 groups = c("0-10", "10-20", "20-30", "30-40", "40-50", "50-60"),
 proportions = c(0.1, 0.2, 0.3, 0.4, 0.2, 0.1)) \%>\%
 # make sure col names match survey dataset
 dplyr::rename(age_grp = groups,
 sex = strata,
 population = n)

# define a fake dataset of cluster listings
# including cluster names and number of households
cz <- tibble::tribble(
  ~cluster, ~n_houses,
  "Village A",        23,
  "Village B",        42,
  "Village C",        56,
  "Village D",        38
)


# add weights to a stratified simple random sample
# weight based on age group and sex
add_weights(x, p = p, age_grp, sex,
population = population, method = "stratified")

# add weights to a cluster sample
# include weights for cluster, household and individual levels
add_weights(x, cz = cz,
cluster_cz = cluster, household_cz = n_houses,
cluster_x = cluster, household_x = household_id,
individuals_eligible_x = eligibile_n, individuals_interviewed_x = surveyed_n,
ignore_cluster = FALSE, ignore_household = FALSE,
method = "cluster")


# add weights to a cluster sample
# ignore weights for cluster and household level (set equal to 1)
# only include weights at individual level
add_weights(x, cz = cz,
cluster_cz = cluster, household_cz = n_houses,
cluster_x = cluster, household_x = household_id,
individuals_eligible_x = eligibile_n, individuals_interviewed_x = surveyed_n,
ignore_cluster = TRUE, ignore_household = TRUE,
method = "cluster")

}
\author{
Zhian N. Kamvar Alex Spina Lukas Richter
}
