% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-categories.R
\name{age_categories}
\alias{age_categories}
\title{Create an age group variable
either give distinct breaks or specify vals for sequence}
\usage{
age_categories(x, breakers = NA, lower = 0, upper = NA, by = 10,
  separator = "-", above.char = "+")
}
\description{
Create an age group variable
either give distinct breaks or specify vals for sequence
}
