% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apyramid_exports.R
\docType{data}
\name{age_pyramid}
\alias{age_pyramid}
\title{Functions re-exported from {apyramid}}
\format{
An object of class \code{function} of length 1.
}
\usage{
age_pyramid
}
\description{
Functions re-exported from {apyramid}
}
\seealso{
\pkg{apyramid} functions:
\itemize{
\item \code{\link[apyramid:age_pyramid]{apyramid::age_pyramid()}}: 	Plot a population pyramid (age-sex) from a dataframe
}
}
\keyword{datasets}
