% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/block_shape.R
\docType{data}
\name{block_shape}
\alias{block_shape}
\title{Quartier shapefile for example outbreak Hepatatis E virus or Acute Jaundice Syndrome (AJS)}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0188240}{Article}
}
\usage{
data(block_shape)
}
\description{
Shapefile for blocks (polygons) from a real outbreak of Hepatitis E virus (HEV) infection which
occurred in the Chadian town of Am Timan between October 2016 and
April 2017. The Chadian Ministry of Health (MoH) has approved the use
of this data for training purposes.
Please note, that some data has been adapted in order to best achieve
training objectives. Also, the GPS coordinates included in the dataset
do not correspond to real cases identified during this outbreak.
They have been generated exclusively for training purposes.
}
\examples{
data(block_shape)
head(block_shape)
}
\references{
None
(\href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0188240}{PLoS})
}
\keyword{datasets}
