% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_frame_from_2x2.R
\name{data_frame_from_2x2}
\alias{data_frame_from_2x2}
\title{create a data frame from a 2x2 matrix}
\usage{
data_frame_from_2x2(x)
}
\arguments{
\item{x}{a 2x2 matrix or 3D array with exposure variable in rows and outcome
in columns}
}
\value{
a data frame with the important combinations:
\itemize{
\item A_exp_cases
\item B_exp_controls
\item C_unexp_cases
\item D_unexp_controls
\item total_cases (A + B)
\item total_controls (C + D)
\item total_exposed (A + C)
\item total_unexposed (B + D)
\item total (A + B + C + D)
}
}
\description{
create a data frame from a 2x2 matrix
}
\examples{
arr <- c(10, 35, 90, 465, 36, 25, 164, 175)
arr <- array(arr, dim = c(2, 2, 2),
             dimnames = list(risk = c(TRUE, FALSE),
                             outcome = c(TRUE, FALSE),
                             old = c(FALSE, TRUE))
       )
arr
data_frame_from_2x2(arr)
}
