% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptive-table.R
\name{descriptive}
\alias{descriptive}
\title{Counts with proportions and totals table}
\usage{
descriptive(df, counter, grouper = NA, multiplier = 100, digits = 1)
}
\description{
Counts with proportions and totals table
}
