% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datasets.R
\name{download_kobo}
\alias{download_kobo}
\title{Access kobo data dictionary}
\usage{
download_kobo(path = rstudioapi::selectDirectory())
}
\arguments{
\item{path}{\strong{Run with no arguments to pick a directory interactively}. Path on your computer where the file(s) should be saved to \emph{string}}
}
\description{
Access kobo data dictionary
}
