% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datasets.R
\name{download_outbreak_linelist}
\alias{download_outbreak_linelist}
\title{Access AJS outbreak linelist}
\usage{
download_outbreak_linelist(path = rstudioapi::selectDirectory())
}
\arguments{
\item{path}{\strong{Run with no arguments to pick a directory interactively}. Path on your computer where the file(s) should be saved to \emph{string}}
}
\description{
Access AJS outbreak linelist
}
