% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datasets.R
\name{download_shape_block}
\alias{download_shape_block}
\title{Access Am Timan blocks shapefile}
\usage{
download_shape_block(path = rstudioapi::selectDirectory())
}
\arguments{
\item{path}{\strong{Run with no arguments to pick a directory interactively}. Path on your computer where the file(s) should be saved to \emph{string}}
}
\description{
Access Am Timan blocks shapefile
}
