% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_survey.R
\name{download_survey}
\alias{download_survey}
\title{download_survey

Access fake mortality survey data}
\usage{
download_survey(path = rstudioapi::selectDirectory())
}
\arguments{
\item{path}{\strong{Run with no arguments to pick a directory interactively}. Path on your computer where the file(s) should be saved to \emph{string}}
}
\description{
download_survey

Access fake mortality survey data
}
