% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidict_exports.R
\name{dict_rename_helper}
\alias{dict_rename_helper}
\alias{gen_data}
\alias{msf_dict}
\alias{msf_dict_rename_helper}
\alias{read_dict}
\title{Functions re-expored from {epidict}}
\usage{
dict_rename_helper(
  dictionary,
  varnames,
  varnames_type,
  rmd,
  copy_to_clipboard = TRUE
)

gen_data(dictionary, varnames = "name", numcases = 300, org = "MSF")

msf_dict(dictionary, tibble = TRUE, long = TRUE, compact = TRUE)

msf_dict_rename_helper(dictionary, copy_to_clipboard = TRUE)

read_dict(path, sheet, format, tibble = TRUE, long = TRUE, compact = TRUE)
}
\description{
Functions re-expored from {epidict}
}
\seealso{
Dictionaries: \code{\link[epidict:msf_dict]{epidict::msf_dict()}}\cr
Renaming: \code{\link[epidict:msf_dict_rename_helper]{epidict::msf_dict_rename_helper()}}\cr
Generator: \code{\link[epidict:gen_data]{epidict::gen_data()}}
}
