% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epikit_exports.R
\docType{import}
\name{add_weights_cluster}
\alias{add_weights_cluster}
\alias{add_weights_strata}
\alias{age_categories}
\alias{assert_positive_timespan}
\alias{augment_redundant}
\alias{constrain_dates}
\alias{dots_to_charlist}
\alias{fac_from_num}
\alias{find_breaks}
\alias{find_date_cause}
\alias{find_end_date}
\alias{find_start_date}
\alias{fmt_ci}
\alias{fmt_ci_df}
\alias{fmt_count}
\alias{fmt_pci}
\alias{fmt_pci_df}
\alias{gen_polygon}
\alias{gen_population}
\alias{group_age_categories}
\alias{merge_ci_df}
\alias{merge_pci_df}
\alias{rename_redundant}
\alias{unite_ci}
\alias{zcurve}
\title{Exported functions from epikit}
\seealso{
\code{\link[epikit:add_weights_cluster]{epikit::add_weights_cluster()}}, \code{\link[epikit:add_weights_strata]{epikit::add_weights_strata()}},
\code{\link[epikit:age_categories]{epikit::age_categories()}}, \code{\link[epikit:find_date_cause]{epikit::assert_positive_timespan()}},
\code{\link[epikit:rename_redundant]{epikit::augment_redundant()}},
\code{\link[epikit:find_date_cause]{epikit::constrain_dates()}}, \code{\link[epikit:dots_to_charlist]{epikit::dots_to_charlist()}},
\code{\link[epikit:fac_from_num]{epikit::fac_from_num()}}, \code{\link[epikit:find_breaks]{epikit::find_breaks()}},
\code{\link[epikit:find_date_cause]{epikit::find_date_cause()}}, \code{\link[epikit:find_date_cause]{epikit::find_end_date()}},
\code{\link[epikit:find_date_cause]{epikit::find_start_date()}}, \code{\link[epikit:fmt_ci]{epikit::fmt_ci()}},
\code{\link[epikit:fmt_ci]{epikit::fmt_ci_df()}}, \code{\link[epikit:fmt_count]{epikit::fmt_count()}}, \code{\link[epikit:fmt_ci]{epikit::fmt_pci()}},
\code{\link[epikit:fmt_ci]{epikit::fmt_pci_df()}}, \code{\link[epikit:age_categories]{epikit::group_age_categories()}},
\code{\link[epikit:unite_ci]{epikit::merge_ci_df()}}, \code{\link[epikit:unite_ci]{epikit::merge_pci_df()}},
\code{\link[epikit:rename_redundant]{epikit::rename_redundant()}},
\code{\link[epikit:unite_ci]{epikit::unite_ci()}}, \code{\link[epikit:zcurve]{epikit::zcurve()}}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{epikit}{\code{\link[epikit]{add_weights_cluster}}, \code{\link[epikit]{add_weights_strata}}, \code{\link[epikit]{age_categories}}, \code{\link[epikit:find_date_cause]{assert_positive_timespan}}, \code{\link[epikit:rename_redundant]{augment_redundant}}, \code{\link[epikit:find_date_cause]{constrain_dates}}, \code{\link[epikit]{dots_to_charlist}}, \code{\link[epikit]{fac_from_num}}, \code{\link[epikit]{find_breaks}}, \code{\link[epikit]{find_date_cause}}, \code{\link[epikit:find_date_cause]{find_end_date}}, \code{\link[epikit:find_date_cause]{find_start_date}}, \code{\link[epikit]{fmt_ci}}, \code{\link[epikit:fmt_ci]{fmt_ci_df}}, \code{\link[epikit]{fmt_count}}, \code{\link[epikit:fmt_ci]{fmt_pci}}, \code{\link[epikit:fmt_ci]{fmt_pci_df}}, \code{\link[epikit]{gen_polygon}}, \code{\link[epikit]{gen_population}}, \code{\link[epikit:age_categories]{group_age_categories}}, \code{\link[epikit:unite_ci]{merge_ci_df}}, \code{\link[epikit:unite_ci]{merge_pci_df}}, \code{\link[epikit]{rename_redundant}}, \code{\link[epikit]{unite_ci}}, \code{\link[epikit]{zcurve}}}
}}

