% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitabulate_exports.R
\docType{import}
\name{case_fatality_rate}
\alias{case_fatality_rate}
\alias{case_fatality_rate_df}
\alias{mortality_rate}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{epitabulate}{\code{\link[epitabulate:attack_rate]{case_fatality_rate}}, \code{\link[epitabulate:attack_rate]{case_fatality_rate_df}}, \code{\link[epitabulate:attack_rate]{mortality_rate}}}
}}

