% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_date_cause.R
\name{find_date_cause}
\alias{find_date_cause}
\alias{find_start_date}
\alias{find_end_date}
\title{Find the first date beyond a cutoff in several columns}
\usage{
find_date_cause(x, ..., datecol = "start_date",
  datereason = "start_date_reason", cutoff = NA, period = "after")

find_start_date(x, ..., datecol = "start_date",
  datereason = "start_date_reason", cutoff = NA)

find_end_date(x, ..., datecol = "end_date",
  datereason = "end_date_reason", cutoff = NA)
}
\arguments{
\item{x}{a data frame}

\item{...}{an ordered series of date columns (i.e. the most important date
to be considered first).}

\item{datecol}{the name of the new column to contain the dates}

\item{datereason}{the name of the column to contain the name of the column
from which the date came.}

\item{cutoff}{the cutoff date}

\item{period}{one of either "before" or "after" indicating that the new
column should only contain dates before or after the cutoff date.}
}
\description{
This function will find the first date in an orderd series of columns that
is either before or after a cutoff date, inclusive.
}
