% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_fun.R
\name{fmt_ci}
\alias{fmt_ci}
\alias{fmt_pci}
\alias{fmt_pci_df}
\alias{fmt_ci_df}
\alias{merge_ci_df}
\title{Helper to format confidence interval for text}
\usage{
fmt_ci(e = numeric(), l = numeric(), u = numeric(), digits = 2)

fmt_pci(e = numeric(), l = numeric(), u = numeric(), digits = 2)

fmt_pci_df(x, e = 3, l = e + 1, u = e + 2, digits = 2)

fmt_ci_df(x, e = 3, l = e + 1, u = e + 2, digits = 2)

merge_ci_df(x, e = 3, l = e + 1, u = e + 2, digits = 2)
}
\arguments{
\item{e}{the column of the estimate (defaults to the third column). Otherwise, a number}

\item{l}{the column of the lower bound (defaults to the fourth column). Otherwise, a number}

\item{u}{the column of the upper bound (defaults to the fifth column), otherwise, a number}

\item{digits}{the number of digits to show}

\item{x}{a data frame}
}
\value{
a text string in the format of "e\% (CI l--u)"
}
\description{
This function is mainly used for placing in the text fields of Rmarkdown
reports. You can use it by writing it in something like this:
\preformatted{The CFR for Bamako is `r fmt_pci(case_fatality_rate(10, 50))`}
which will render like this: "The CFR for Bamako is 20.00\% (CI 11.24--33.04)"
}
\examples{

cfr <- case_fatality_rate(10, 50)
fmt_ci_df(cfr)
fmt_ci_df(cfr, d = 1)
# If the data starts at a different column, specify a different number
fmt_ci_df(cfr[-1], 2, d = 1)

# It's also possible to provide numbers directly
fmt_ci(pi, pi - runif(1), pi + runif(1))
}
