% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inline_fun.R
\name{fmt_count}
\alias{fmt_count}
\title{Counts and proportions inline}
\usage{
fmt_count(x, ...)
}
\arguments{
\item{x}{a data frame}

\item{...}{an expression or series of expressions to pass to \code{\link[dplyr:filter]{dplyr::filter()}}}
}
\description{
These functions will give proportions for different variables inline.
}
\examples{

fmt_count(mtcars, cyl > 3, hp < 100)
fmt_count(iris, Species == "virginica")
}
