% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_polygon.R
\name{gen_polygon}
\alias{gen_polygon}
\title{Fake spatial data as polygons
This function returns a polygon which is split in to regions based on a
supplied vector of names}
\usage{
gen_polygon(regions)
}
\arguments{
\item{regions}{A string of names for each region to label the polygon with}
}
\description{
Fake spatial data as polygons
This function returns a polygon which is split in to regions based on a
supplied vector of names
}
\references{
The coordinates used for the polygon are of Vienna, Austria.
based off government data (see \href{https://www.data.gv.at/katalog/dataset/stadt-wien_bezirksgrenzenwien}{metadata})
}
