% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_population.R
\name{gen_population}
\alias{gen_population}
\title{Generate population counts from estimated population age breakdowns
This generates based on predefined age groups and proportions,
however you could also define these yourself.}
\usage{
gen_population(total_pop = 1000, groups = c("0-4", "5-14", "15-24",
  "25-34", "35-44", "45+"), proportions = c(0.182, 0.278, 0.26, 0.11,
  0.07, 0.1), tibble = TRUE)
}
\arguments{
\item{total_pop}{The overal population count of interest - the default is 1000 people}

\item{proportions}{A numeric vector specifying the proportions for each age group -
the default is c(0.182, 0.278, 0.26, 0.11, 0.07, 0.10)}

\item{tibble}{Return data as a tidyverse tibble (default is TRUE)}

\item{age_groups}{A character vector of groups -
the default is set for age groups: c("0-4","5-14","15-24","25-34","35-44","45+")}
}
\description{
Generate population counts from estimated population age breakdowns
This generates based on predefined age groups and proportions,
however you could also define these yourself.
}
