% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_datasets.R
\name{generic_download}
\alias{generic_download}
\title{Internal function (not exported - i.e. not userfacing) to reduce code duplication}
\usage{
generic_download(path = rstudioapi::selectDirectory(), dataset)
}
\arguments{
\item{path}{\strong{Run with no arguments to pick a directory interactively}. Path on your computer where the file(s) should be saved to \emph{string}}

\item{dataset}{The name of the file to save}
}
\description{
Internal function (not exported - i.e. not userfacing) to reduce code duplication
}
