% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitabulate_exports.R
\docType{import}
\name{gt_mh_odds}
\alias{gt_mh_odds}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{epitabulate}{\code{\link[epitabulate:gtsummary_wrappers]{gt_mh_odds}}}
}}

