% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kobo_vars.R
\docType{data}
\name{kobo_vars}
\alias{kobo_vars}
\title{Example Kobo data dictionary, showing variables}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0188240}{Article}
}
\usage{
data(kobo_vars)
}
\description{
This is an example of a Kobo data dictionary (the sheet showing information on the variables)
}
\examples{
data(kobo_vars)
head(kobo_vars)
}
\references{
None
(\href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0188240}{PLoS})
}
\keyword{datasets}
