% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidict_exports.R
\name{msf_dict}
\alias{msf_dict}
\alias{msf_dict_rename_helper}
\alias{gen_data}
\title{Functions re-expored from {epidict}}
\usage{
msf_dict(disease, tibble = TRUE, compact = TRUE, long = TRUE)

msf_dict_rename_helper(dictionary, copy_to_clipboard = TRUE)

gen_data(dictionary, varnames = "name", numcases = 300, org = "MSF")
}
\description{
Functions re-expored from {epidict}
}
\seealso{
Dictionaries: \code{\link[epidict:msf_dict]{epidict::msf_dict()}}\cr
Renaming: \code{\link[epidict:msf_dict_rename_helper]{epidict::msf_dict_rename_helper()}}\cr
Generator: \code{\link[epidict:gen_data]{epidict::gen_data()}}
}
