% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep-package.R
\docType{package}
\name{sitrep-package}
\alias{sitrep}
\alias{sitrep-package}
\title{sitrep: Report templates and helper functions for EPI}
\description{
Report templates and helper functions for EPI.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r4epis.netlify.com}
  \item \url{https://github.com/R4EPI/sitrep}
  \item Report bugs at \url{https://github.com/R4EPI/sitrep/issues}
}

}
\author{
\strong{Maintainer}: Zhian N. Kamvar \email{zkamvar@gmail.com}

Authors:
\itemize{
  \item Dirk Schumacher \email{mai@dirk-schumacher.net}
  \item Alex Spina
  \item Kate Doyle
  \item Chris Jarvis
  \item Lukas Richter
}

Other contributors:
\itemize{
  \item Annick Lenglet [copyright holder]
  \item Neale Batra [contributor]
}

}
\keyword{internal}
