% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nutrition.R
\name{stratify_zscores}
\alias{stratify_zscores}
\title{Stratifies z-scores by variables in the survey data}
\usage{
stratify_zscores(survey_data, by, ...)
}
\arguments{
\item{survey_data}{a data.frame with the survey data}

\item{by}{a column by which you want to stratify the data}

\item{...}{columns having zscore indicators}
}
\description{
Stratifies z-scores by variables in the survey data
}
\examples{
# A crude example, but it works
stratify_zscores(mtcars, by = "gear", am, vs)
}
