% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey_ciprop_foreach_binary_variable}
\alias{survey_ciprop_foreach_binary_variable}
\title{Compute summary statistics for a list of binary variables}
\usage{
survey_ciprop_foreach_binary_variable(design, variables)
}
\arguments{
\item{design}{a design object from the \code{\link{survey}} package}

\item{variables}{a character vector of variables present in the design}
}
\description{
Compute summary statistics for a list of binary variables
}
