% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_data.R
\docType{data}
\name{survey_data}
\alias{survey_data}
\title{Example data from a retrospective mortality survey (linelist)}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0188240}{Article}
}
\usage{
data(survey_data)
}
\description{
This is an example of data from a retrospective mortality survey
}
\examples{
data(survey_data)
head(survey_data)
}
\references{
None
(\href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0188240}{PLoS})
}
\keyword{datasets}
