% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_hhdata.R
\docType{data}
\name{survey_hhdata}
\alias{survey_hhdata}
\title{Example data from a retrospective mortality survey (household members)}
\format{
An object of class \code{"cross"}; see \code{\link[qtl]{read.cross}}.
}
\source{
\href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0188240}{Article}
}
\usage{
data(survey_hhdata)
}
\description{
This is data from a retrospective mortality survey, showing data from household members
}
\examples{
data(survey_hhdata)
head(survey_hhdata)
}
\references{
None
(\href{https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0188240}{PLoS})
}
\keyword{datasets}
