% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{suvery_ciprob}
\alias{suvery_ciprob}
\title{Compute summary statistics for a single binary variable}
\usage{
suvery_ciprob(design, variable)
}
\arguments{
\item{design}{a design object from the \code{\link{survey}} package}

\item{variable}{a variable name}
}
\description{
Compute summary statistics for a single binary variable
}
