% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{suvery_ciprob_stratified}
\alias{suvery_ciprob_stratified}
\title{Compute stratified summary statistics for a binary variable}
\usage{
suvery_ciprob_stratified(design, variable, by)
}
\arguments{
\item{design}{a design object from the \code{\link{survey}} package}

\item{variable}{a variable name}

\item{by}{another variable used for stratification}
}
\description{
Compute stratified summary statistics for a binary variable
}
