% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epibuffet_exports.R
\docType{data}
\name{tab_survey}
\alias{tab_survey}
\alias{tab_linelist}
\alias{tab_univariate}
\alias{data_frame_from_2x2}
\title{Functions re-expored from {epibuffet}}
\format{An object of class \code{function} of length 1.}
\usage{
tab_survey

tab_linelist

tab_univariate

data_frame_from_2x2
}
\description{
Functions re-expored from {epibuffet}
}
\seealso{
\pkg{epibuffet} functions:
\itemize{
\item \code{\link[epibuffet:tab_linelist]{epibuffet::tab_linelist()}}: tabulate linelist data
\item \code{\link[epibuffet:tab_survey]{epibuffet::tab_survey()}}: tabulate survey data
\item \code{\link[epibuffet:tab_univariate]{epibuffet::tab_univariate()}}: caluclate odds, risk, and incidence risk
ratios for multiple variables from linelist data.
\item \code{\link[epibuffet:data_frame_from_2x2]{epibuffet::data_frame_from_2x2()}}: convert a 2x2(x2) table to a data
frame clearly labelling the (un)exposed (non)case combinations and their
totals.
}
}
\keyword{datasets}
