% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_survey.R
\name{tabulate_survey}
\alias{tabulate_survey}
\alias{tabulate_binary_survey}
\title{Tabulate survey design objects by a categorical and another stratifying variable}
\usage{
tabulate_survey(x, var, strata = NULL, pretty = TRUE, digits = 1)

tabulate_binary_survey(x, ..., keep = NULL, invert = FALSE,
  pretty = TRUE, digits = 1)
}
\arguments{
\item{x}{a survey design object}

\item{var}{the bare name of a categorical variable}

\item{strata}{a variable to stratify the results by}

\item{pretty}{if \code{TRUE}, default, the proportion and CI are merged}

\item{digits}{if \code{prittey = FALSE}, this indicates the number of digits used
for proportion and CI}

\item{...}{binary variables for tabulation}

\item{keep}{a vector of binary values to keep}

\item{invert}{if \code{TRUE}, the kept values are rejected. Defaults to \code{FALSE}}
}
\value{
a tibble
}
\description{
Tabulate survey design objects by a categorical and another stratifying variable
}
\examples{
library(srvyr)
library(survey)
data(api)

# stratified sample
apistrat \%>\%
  as_survey_design(strata = stype, weights = pw) \%>\%
  tabulate_survey(stype, awards)

apistrat \%>\%
  as_survey_design(strata = stype, weights = pw) \%>\%
  tabulate_binary_survey(stype, awards, keep = c("Yes", "E"))

apistrat \%>\%
  as_survey_design(strata = stype, weights = pw) \%>\%
  tabulate_binary_survey(stype, awards, keep = c("Yes", "E"), invert = TRUE)
}
