% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unite_ci.R
\name{unite_ci}
\alias{unite_ci}
\title{Unite estimates and confidence intervals}
\usage{
unite_ci(
  x,
  col = NULL,
  ...,
  remove = TRUE,
  digits = 2,
  m100 = TRUE,
  percent = FALSE,
  ci = FALSE
)
}
\arguments{
\item{x}{a data frame with at least three columns defining an estimate, lower
bounds, and upper bounds.}

\item{col}{the quoted name of the replacement column to create}

\item{...}{three columns to bind together in the order of Estimate, Lower, and
Upper.}

\item{remove}{if \code{TRUE} (default), the three columns in \code{...} will be replaced by \code{col}}

\item{digits}{the number of digits to retain for the confidence interval.}

\item{m100}{\code{TRUE} if the result should be multiplied by 100}

\item{percent}{\code{TRUE} if the result should have a percent symbol added.}

\item{ci}{\code{TRUE} if the result should include "CI" within the braces (defaults to FALSE)}
}
\description{
create a character column by combining estimate, lower and upper columns.
This is similar to \code{\link[tidyr:unite]{tidyr::unite()}}.
}
\examples{

print(cfr <- case_fatality_rate((1:4)*10, 50))
unite_ci(cfr, "CFR (CI)", cfr, lower, upper, m100 = FALSE, percent = TRUE)

}
