% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate-analysis.R
\name{univariate_analysis}
\alias{univariate_analysis}
\title{# Inspired by Daniel Gardiner
https://github.com/DanielGardiner/UsefulFunctions/blob/efffde624d424d977651ed1a9ee4430cbf2b0d6f/single.variable.analysis.v0.3.R#L12}
\usage{
univariate_analysis(outcome, ...)
}
\description{
# Inspired by Daniel Gardiner
https://github.com/DanielGardiner/UsefulFunctions/blob/efffde624d424d977651ed1a9ee4430cbf2b0d6f/single.variable.analysis.v0.3.R#L12
}
