% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcurve.R
\name{zcurve}
\alias{zcurve}
\title{Create a curve comparing observed Z-scores to the WHO standard.}
\usage{
zcurve(zscore)
}
\arguments{
\item{zscore}{a umeric vector containing computed zscores}
}
\value{
a ggplot2 object that is customisable via the ggplot2 package.
}
\description{
Create a curve comparing observed Z-scores to the WHO standard.
}
\examples{
library("ggplot2")
dat <- rnorm(204) + runif(1) # slightly skewed
zcurve(dat) +
  labs(title = "Weight-for-Height Z-scores") +
  theme_classic()
}
