% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_weights.R
\name{add_weights}
\alias{add_weights}
\title{Add a column of weights to a data frame}
\usage{
add_weights(x, p, ..., population = population,
  surv_weight = surv_weight, surv_weight_ID = surv_weight_ID)
}
\arguments{
\item{x}{a data frame of survey data}

\item{p}{a data frame containing pouplation data for groups in \code{...}}

\item{...}{shared grouping columns across both \code{x} and \code{p}. These are used
to match the weights to the correct subset of the pouplation.}

\item{population}{the column in \code{p} that defines the pouplation numbers}

\item{surv_weight}{the name of the new column to store the weights. Defaults to
"surv_weight".}

\item{surv_weight_ID}{the name of the new ID column to be created. Defaults to
"surv_weight_ID"}
}
\description{
Add a column of weights to a data frame
}
\author{
Zhian N. Kamvar Alex Spina Lukas Richter
}
