% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-categories.R
\name{age_categories}
\alias{age_categories}
\title{Create an age group variable
either give distinct breaks or specify vals for sequence}
\usage{
age_categories(x, breakers = NA, lower = 0, upper = NA, by = 10,
  separator = "-", above.char = "+")
}
\arguments{
\item{x}{Your age variable}

\item{breakers}{A string. Age category breaks you can define within c(). Alternatively use "lower",
"upper" and "by" to set these breaks based on a sequence.}

\item{lower}{A number. The lowest age value you want to consider (default is 0)}

\item{upper}{A number. The highest age value you want to consider}

\item{by}{A number. The number of years you want between groups}

\item{separator}{A character that you want to have between ages in group names. The default is "-" producing e.g. 0-10.}

\item{above.char}{A character that you want to have after your highest age group. The default is "+" producing e.g. 80+}
}
\description{
Create an age group variable
either give distinct breaks or specify vals for sequence
}
