% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_outbreak_pop.R
\name{download_outbreak_pop}
\alias{download_outbreak_pop}
\title{download_outbreak_pop

Access AJS outbreak population data}
\usage{
download_outbreak_pop(path = rstudioapi::selectDirectory())
}
\arguments{
\item{path}{\strong{Run with no arguments to pick a directory interactively}. Path on your computer where the file(s) should be saved to \emph{string}}
}
\description{
download_outbreak_pop

Access AJS outbreak population data
}
