% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_shape_quartier.R
\name{download_shape_quartier}
\alias{download_shape_quartier}
\title{download_shape_quartier

Access Am Timan quartier shapefile}
\usage{
download_shape_quartier(path = rstudioapi::selectDirectory())
}
\arguments{
\item{path}{\strong{Run with no arguments to pick a directory interactively}. Path on your computer where the file(s) should be saved to \emph{string}}
}
\description{
download_shape_quartier

Access Am Timan quartier shapefile
}
