% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epikit_exports.R
\name{add_weights_cluster}
\alias{add_weights_cluster}
\alias{add_weights_strata}
\alias{age_categories}
\alias{assert_positive_timespan}
\alias{augment_redundant}
\alias{constrain_dates}
\alias{dots_to_charlist}
\alias{fac_from_num}
\alias{find_breaks}
\alias{find_date_cause}
\alias{find_end_date}
\alias{find_start_date}
\alias{fmt_ci}
\alias{fmt_ci_df}
\alias{fmt_count}
\alias{fmt_pci}
\alias{fmt_pci_df}
\alias{gen_polygon}
\alias{gen_population}
\alias{group_age_categories}
\alias{merge_ci_df}
\alias{merge_pci_df}
\alias{rename_redundant}
\alias{unite_ci}
\alias{zcurve}
\title{Exported functions from epikit}
\usage{
add_weights_cluster(
  x,
  cl,
  eligible,
  interviewed,
  cluster_x = NULL,
  cluster_cl = NULL,
  household_x = NULL,
  household_cl = NULL,
  ignore_cluster = TRUE,
  ignore_household = TRUE,
  surv_weight = "surv_weight",
  surv_weight_ID = "surv_weight_ID"
)

add_weights_strata(
  x,
  p,
  ...,
  population = population,
  surv_weight = "surv_weight",
  surv_weight_ID = "surv_weight_ID"
)

age_categories(
  x,
  breakers = NULL,
  lower = 0,
  upper = NULL,
  by = 10,
  separator = "-",
  ceiling = FALSE,
  above.char = "+"
)

assert_positive_timespan(x, date_start, date_end)

augment_redundant(x, ...)

constrain_dates(i, period_start, period_end, boundary = "both")

dots_to_charlist(parent = 1L)

fac_from_num(x)

find_breaks(n, breaks = 4, snap = 1, ceiling = FALSE)

find_date_cause(
  x,
  ...,
  period_start = NULL,
  period_end = NULL,
  datecol = "start_date",
  datereason = "start_date_reason",
  na_fill = "start"
)

find_end_date(
  x,
  ...,
  period_start = NULL,
  period_end = NULL,
  datecol = "end_date",
  datereason = "end_date_reason"
)

find_start_date(
  x,
  ...,
  period_start = NULL,
  period_end = NULL,
  datecol = "start_date",
  datereason = "start_date_reason"
)

fmt_ci(
  e = numeric(),
  l = numeric(),
  u = numeric(),
  digits = 2,
  percent = TRUE,
  separator = "-"
)

fmt_ci_df(
  x,
  e = 3,
  l = e + 1,
  u = e + 2,
  digits = 2,
  percent = TRUE,
  separator = "-"
)

fmt_count(x, ...)

fmt_pci(
  e = numeric(),
  l = numeric(),
  u = numeric(),
  digits = 2,
  percent = TRUE,
  separator = "-"
)

fmt_pci_df(
  x,
  e = 3,
  l = e + 1,
  u = e + 2,
  digits = 2,
  percent = TRUE,
  separator = "-"
)

gen_polygon(regions)

gen_population(
  total_pop = 1000,
  groups = c("0-4", "5-14", "15-29", "30-44", "45+"),
  strata = c("Male", "Female"),
  proportions = c(0.079, 0.134, 0.139, 0.082, 0.066),
  counts = NULL,
  tibble = TRUE
)

group_age_categories(
  dat,
  years = NULL,
  months = NULL,
  weeks = NULL,
  days = NULL,
  one_column = TRUE,
  drop_empty_overlaps = TRUE
)

merge_ci_df(x, e = 3, l = e + 1, u = e + 2, digits = 2, separator = "-")

merge_pci_df(x, e = 3, l = e + 1, u = e + 2, digits = 2, separator = "-")

rename_redundant(x, ...)

unite_ci(
  x,
  col = NULL,
  ...,
  remove = TRUE,
  digits = 2,
  m100 = TRUE,
  percent = FALSE,
  ci = FALSE,
  separator = "-"
)

zcurve(x, zscore)
}
\description{
Exported functions from epikit
}
\seealso{
\code{\link[epikit:add_weights_cluster]{epikit::add_weights_cluster()}}, \code{\link[epikit:add_weights_strata]{epikit::add_weights_strata()}},
\code{\link[epikit:age_categories]{epikit::age_categories()}}, \code{\link[epikit:find_date_cause]{epikit::assert_positive_timespan()}},
\code{\link[epikit:rename_redundant]{epikit::augment_redundant()}},
\code{\link[epikit:find_date_cause]{epikit::constrain_dates()}}, \code{\link[epikit:dots_to_charlist]{epikit::dots_to_charlist()}},
\code{\link[epikit:fac_from_num]{epikit::fac_from_num()}}, \code{\link[epikit:find_breaks]{epikit::find_breaks()}},
\code{\link[epikit:find_date_cause]{epikit::find_date_cause()}}, \code{\link[epikit:find_date_cause]{epikit::find_end_date()}},
\code{\link[epikit:find_date_cause]{epikit::find_start_date()}}, \code{\link[epikit:fmt_ci]{epikit::fmt_ci()}},
\code{\link[epikit:fmt_ci]{epikit::fmt_ci_df()}}, \code{\link[epikit:fmt_count]{epikit::fmt_count()}}, \code{\link[epikit:fmt_ci]{epikit::fmt_pci()}},
\code{\link[epikit:fmt_ci]{epikit::fmt_pci_df()}}, \code{\link[epikit:age_categories]{epikit::group_age_categories()}},
\code{\link[epikit:unite_ci]{epikit::merge_ci_df()}}, \code{\link[epikit:unite_ci]{epikit::merge_pci_df()}},
\code{\link[epikit:rename_redundant]{epikit::rename_redundant()}},
\code{\link[epikit:unite_ci]{epikit::unite_ci()}}, \code{\link[epikit:zcurve]{epikit::zcurve()}}
}
