% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epikit_exports.R
\docType{data}
\name{age_categories}
\alias{age_categories}
\alias{attack_rate}
\alias{augment_redundant}
\alias{case_fatality_rate}
\alias{case_fatality_rate_df}
\alias{dots_to_charlist}
\alias{fac_from_num}
\alias{find_breaks}
\alias{fmt_ci}
\alias{fmt_ci_df}
\alias{fmt_count}
\alias{fmt_pci}
\alias{fmt_pci_df}
\alias{group_age_categories}
\alias{merge_ci_df}
\alias{merge_pci_df}
\alias{mortality_rate}
\alias{rename_redundant}
\alias{unite_ci}
\title{Exported functions from epikit}
\format{An object of class \code{function} of length 1.}
\usage{
age_categories

attack_rate

augment_redundant

case_fatality_rate

case_fatality_rate_df

dots_to_charlist

fac_from_num

find_breaks

fmt_ci

fmt_ci_df

fmt_count

fmt_pci

fmt_pci_df

group_age_categories

merge_ci_df

merge_pci_df

mortality_rate

rename_redundant

unite_ci
}
\description{
Exported functions from epikit
}
\seealso{
\code{\link[epikit:age_categories]{epikit::age_categories()}}, \code{\link[epikit:attack_rate]{epikit::attack_rate()}},
\code{\link[epikit:augment_redundant]{epikit::augment_redundant()}}, \code{\link[epikit:case_fatality_rate]{epikit::case_fatality_rate()}},
\code{\link[epikit:case_fatality_rate_df]{epikit::case_fatality_rate_df()}}, \code{\link[epikit:dots_to_charlist]{epikit::dots_to_charlist()}},
\code{\link[epikit:fac_from_num]{epikit::fac_from_num()}}, \code{\link[epikit:find_breaks]{epikit::find_breaks()}}, \code{\link[epikit:fmt_ci]{epikit::fmt_ci()}},
\code{\link[epikit:fmt_ci_df]{epikit::fmt_ci_df()}}, \code{\link[epikit:fmt_count]{epikit::fmt_count()}}, \code{\link[epikit:fmt_pci]{epikit::fmt_pci()}},
\code{\link[epikit:fmt_pci_df]{epikit::fmt_pci_df()}}, \code{\link[epikit:group_age_categories]{epikit::group_age_categories()}},
\code{\link[epikit:merge_ci_df]{epikit::merge_ci_df()}}, \code{\link[epikit:merge_pci_df]{epikit::merge_pci_df()}},
\code{\link[epikit:mortality_rate]{epikit::mortality_rate()}},
\code{\link[epikit:rename_redundant]{epikit::rename_redundant()}}, \code{\link[epikit:unite_ci]{epikit::unite_ci()}}
}
\keyword{datasets}
