% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epikit_exports.R
\docType{data}
\name{add_weights_cluster}
\alias{add_weights_cluster}
\alias{add_weights_strata}
\alias{age_categories}
\alias{assert_positive_timespan}
\alias{attack_rate}
\alias{augment_redundant}
\alias{case_fatality_rate}
\alias{case_fatality_rate_df}
\alias{constrain_dates}
\alias{dots_to_charlist}
\alias{fac_from_num}
\alias{find_breaks}
\alias{find_date_cause}
\alias{find_end_date}
\alias{find_start_date}
\alias{fmt_ci}
\alias{fmt_ci_df}
\alias{fmt_count}
\alias{fmt_pci}
\alias{fmt_pci_df}
\alias{gen_polygon}
\alias{gen_population}
\alias{group_age_categories}
\alias{merge_ci_df}
\alias{merge_pci_df}
\alias{mortality_rate}
\alias{rename_redundant}
\alias{unite_ci}
\alias{zcurve}
\title{Exported functions from epikit}
\format{
An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.
}
\usage{
add_weights_cluster

add_weights_strata

age_categories

assert_positive_timespan

attack_rate

augment_redundant

case_fatality_rate

case_fatality_rate_df

constrain_dates

dots_to_charlist

fac_from_num

find_breaks

find_date_cause

find_end_date

find_start_date

fmt_ci

fmt_ci_df

fmt_count

fmt_pci

fmt_pci_df

gen_polygon

gen_population

group_age_categories

merge_ci_df

merge_pci_df

mortality_rate

rename_redundant

unite_ci

zcurve
}
\description{
Exported functions from epikit
}
\seealso{
\code{\link[epikit:add_weights_cluster]{epikit::add_weights_cluster()}}, \code{\link[epikit:add_weights_strata]{epikit::add_weights_strata()}},
\code{\link[epikit:age_categories]{epikit::age_categories()}}, \code{\link[epikit:find_date_cause]{epikit::assert_positive_timespan()}},
\code{\link[epikit:attack_rate]{epikit::attack_rate()}}, \code{\link[epikit:rename_redundant]{epikit::augment_redundant()}},
\code{\link[epikit:attack_rate]{epikit::case_fatality_rate()}}, \code{\link[epikit:attack_rate]{epikit::case_fatality_rate_df()}},
\code{\link[epikit:find_date_cause]{epikit::constrain_dates()}}, \code{\link[epikit:dots_to_charlist]{epikit::dots_to_charlist()}},
\code{\link[epikit:fac_from_num]{epikit::fac_from_num()}}, \code{\link[epikit:find_breaks]{epikit::find_breaks()}},
\code{\link[epikit:find_date_cause]{epikit::find_date_cause()}}, \code{\link[epikit:find_date_cause]{epikit::find_end_date()}},
\code{\link[epikit:find_date_cause]{epikit::find_start_date()}}, \code{\link[epikit:fmt_ci]{epikit::fmt_ci()}},
\code{\link[epikit:fmt_ci]{epikit::fmt_ci_df()}}, \code{\link[epikit:fmt_count]{epikit::fmt_count()}}, \code{\link[epikit:fmt_ci]{epikit::fmt_pci()}},
\code{\link[epikit:fmt_ci]{epikit::fmt_pci_df()}}, \code{\link[epikit:age_categories]{epikit::group_age_categories()}},
\code{\link[epikit:unite_ci]{epikit::merge_ci_df()}}, \code{\link[epikit:unite_ci]{epikit::merge_pci_df()}},
\code{\link[epikit:attack_rate]{epikit::mortality_rate()}},\code{\link[epikit:rename_redundant]{epikit::rename_redundant()}},
\code{\link[epikit:unite_ci]{epikit::unite_ci()}}, \code{\link[epikit:zcurve]{epikit::zcurve()}}
}
\keyword{datasets}
