% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitabulate_exports.R
\name{attack_rate}
\alias{attack_rate}
\alias{case_fatality_rate}
\alias{case_fatality_rate_df}
\alias{mortality_rate}
\alias{add_ar}
\alias{add_cfr}
\alias{add_crosstabs}
\alias{add_mr}
\alias{gt_remove_stat}
\alias{tbl_cmh}
\title{Functions re-exported from {epitabulate}}
\usage{
attack_rate(
  cases,
  population,
  conf_level = 0.95,
  multiplier = 100,
  mergeCI = FALSE,
  digits = 2
)

case_fatality_rate(
  deaths,
  population,
  conf_level = 0.95,
  multiplier = 100,
  mergeCI = FALSE,
  digits = 2
)

case_fatality_rate_df(
  x,
  deaths,
  group = NULL,
  conf_level = 0.95,
  multiplier = 100,
  mergeCI = FALSE,
  digits = 2,
  add_total = FALSE
)

mortality_rate(
  deaths,
  population,
  conf_level = 0.95,
  multiplier = 10^4,
  mergeCI = FALSE,
  digits = 2
)

add_ar(
  gts_object,
  case_var,
  population = NULL,
  multiplier = 10^4,
  drop_tblsummary_stat = FALSE
)

add_cfr(gts_object, deaths_var)

add_crosstabs(x, wide = FALSE)

add_mr(
  gts_object,
  deaths_var,
  population = NULL,
  multiplier = 10^4,
  drop_tblsummary_stat = FALSE
)

gt_remove_stat(gts_object, col_name = "stat_0")

tbl_cmh(
  data,
  case,
  exposure,
  strata,
  measure,
  obstime = NULL,
  conf.level = 0.95
)
}
\description{
Functions re-exported from {epitabulate}
}
\seealso{
\pkg{epitabulate} functions:
\itemize{
\item \code{\link[epitabulate:attack_rate]{epitabulate::attack_rate()}},
\item \code{\link[epitabulate:attack_rate]{epitabulate::case_fatality_rate()}}
\item \code{\link[epitabulate:attack_rate]{epitabulate::case_fatality_rate_df()}},
\item \code{\link[epitabulate:attack_rate]{epitabulate::mortality_rate()}}
\item \code{\link[epitabulate:gt_attackrate]{epitabulate::add_ar()}}: a {gtsummary} wrapper for epitabulate::attack_rate()
\item \code{\link[epitabulate:gt_attackrate]{epitabulate::add_cfr()}}: a {gtsummary} wrapper for
epitabulate::case_fatality_rate()
\item \code{\link[epitabulate:add_crosstabs]{epitabulate::add_crosstabs()}}: a {gtsummary} wrapper to add counts to
a gtsummary::tbl_uvregression()
\item \code{\link[epitabulate:gt_attackrate]{epitabulate::add_mr()}}: a {gtsummary} wrapper for epitabulate::mortality_rate()
\item \code{\link[epitabulate:gtsummary_wrappers]{epitabulate::gt_remove_stat()}}: a {gtsummary} wrapper to remove variables
from a gtsummary table
\item \code{\link[epitabulate:tbl_cmh]{epitabulate::tbl_cmh()}}: a {gtsummary} wrapper to calculate stratified
odds, risk, and incidence risk ratios for multiple variables from linelist data.
}
}
