% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitabulate_exports.R
\docType{import}
\name{attack_rate}
\alias{attack_rate}
\title{Functions re-exported from {epitabulate}}
\seealso{
\pkg{epitabulate} functions:
\itemize{
\item \code{\link[epitabulate:attack_rate]{epitabulate::attack_rate()}},
\item \code{\link[epitabulate:attack_rate]{epitabulate::case_fatality_rate()}}
\item \code{\link[epitabulate:attack_rate]{epitabulate::case_fatality_rate_df()}},
\item \code{\link[epitabulate:attack_rate]{epitabulate::mortality_rate()}}
\item \code{\link[epitabulate:gtsummary_wrappers]{epitabulate::add_ar()}}: a {gtsummary} wrapper for epitabulate::attack_rate()
\item \code{\link[epitabulate:gtsummary_wrappers]{epitabulate::add_cfr()}}: a {gtsummary} wrapper for
epitabulate::case_fatality_rate()
\item \code{\link[epitabulate:gtsummary_wrappers]{epitabulate::add_crosstabs()}}: a {gtsummary} wrapper to add counts to
a gtsummary::tbl_uvregression()
\item \code{\link[epitabulate:gtsummary_wrappers]{epitabulate::add_mr()}}: a {gtsummary} wrapper for epitabulate::mortality_rate()
\item \code{\link[epitabulate:gtsummary_wrappers]{epitabulate::gt_mh_odds()}}: a {gtsummary} wrapper for stratified
univariate regression and mantel-haenszel estimates
\item \code{\link[epitabulate:gtsummary_wrappers]{epitabulate::gt_remove_stat()}}: a {gtsummary} wrapper to remove variables
from a gtsummary table
\item \code{\link[epitabulate:tab_functions]{epitabulate::tab_linelist()}}: tabulate linelist data
\item \code{\link[epitabulate:tab_functions]{epitabulate::tab_survey()}}: tabulate survey data
\item \code{\link[epitabulate:tab_univariate]{epitabulate::tab_univariate()}}: caluclate odds, risk, and incidence risk
ratios for multiple variables from linelist data.
\item \code{\link[epitabulate:data_frame_from_2x2]{epitabulate::data_frame_from_2x2()}}: convert a 2x2(x2) table to a data
frame clearly labelling the (un)exposed (non)case combinations and their
totals.
}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{epitabulate}{\code{\link[epitabulate]{attack_rate}}}
}}

