% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidict_exports.R
\docType{data}
\name{msf_dict}
\alias{msf_dict}
\alias{msf_dict_survey}
\alias{msf_dict_rename_helper}
\alias{gen_data}
\title{Functions re-expored from {epidict}}
\format{
An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.
}
\usage{
msf_dict

msf_dict_survey

msf_dict_rename_helper

gen_data
}
\description{
Functions re-expored from {epidict}
}
\seealso{
Dictionaries: \code{\link[epidict:msf_dict]{epidict::msf_dict()}}, \code{\link[epidict:msf_dict]{epidict::msf_dict_survey()}}\cr
Renaming: \code{\link[epidict:msf_dict_rename_helper]{epidict::msf_dict_rename_helper()}}\cr
Generator: \code{\link[epidict:gen_data]{epidict::gen_data()}}
}
\keyword{datasets}
