% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epidict_exports.R
\docType{import}
\name{msf_dict}
\alias{msf_dict}
\alias{msf_dict_survey}
\alias{msf_dict_rename_helper}
\alias{gen_data}
\title{Functions re-expored from {epidict}}
\seealso{
Dictionaries: \code{\link[epidict:msf_dict]{epidict::msf_dict()}}, \code{\link[epidict:msf_dict]{epidict::msf_dict_survey()}}\cr
Renaming: \code{\link[epidict:msf_dict_rename_helper]{epidict::msf_dict_rename_helper()}}\cr
Generator: \code{\link[epidict:gen_data]{epidict::gen_data()}}
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{epidict}{\code{\link[epidict]{gen_data}}, \code{\link[epidict]{msf_dict}}, \code{\link[epidict]{msf_dict_rename_helper}}, \code{\link[epidict:msf_dict]{msf_dict_survey}}}
}}

