% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msfdict_exports.R
\docType{data}
\name{msf_dict_survey}
\alias{msf_dict_survey}
\alias{msf_dict}
\alias{gen_data}
\title{Functions re-expored from {msfdict}}
\format{An object of class \code{function} of length 1.}
\usage{
msf_dict_survey

msf_dict

gen_data
}
\description{
Functions re-expored from {msfdict}
}
\seealso{
Dictionaries: \code{\link[msfdict:msf_dict]{msfdict::msf_dict()}}, \code{\link[msfdict:msf_dict_survey]{msfdict::msf_dict_survey()}}\cr
Generator: \code{\link[msfdict:gen_data]{msfdict::gen_data()}}
}
\keyword{datasets}
