% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data.R
\name{msf_dict}
\alias{msf_dict}
\title{MSF data dictionaries and dummy datasets
This function reads in MSF data dictioanries and produces randomised datasets based these}
\usage{
msf_dict(disease, name = "MSF-outbreak-dict.xls", tibble = TRUE)
}
\arguments{
\item{disease}{Specify which disease you would like to use.
Currently supports "Cholera", "Measles" and "Meningitis".}

\item{tibble}{Return data dictionary as a tidyverse tibble (default is TRUE)}

\item{varnames}{Specify name of column that contains varnames.
Currently default set to "Item".
(this can probably be deleted once dictionaries standardise)}

\item{numcases}{For fake data, specify the number of cases you want (default is 300)}
}
\description{
MSF data dictionaries and dummy datasets
This function reads in MSF data dictioanries and produces randomised datasets based these
}
