% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age-pyramid.R
\name{plot_age_pyramid}
\alias{plot_age_pyramid}
\title{Plot a population pyramid (age-sex) from a dataframe.}
\usage{
plot_age_pyramid(data, age_group = "age_group", split_by = "sex",
  stack_by = split_by, vertical_lines = FALSE)
}
\arguments{
\item{data}{Your dataframe (e.g. linelist)}

\item{age_group}{the name of a column in the data frame that defines the age
group categories. Defaults to "age_group"}

\item{split_by}{the name of a column in the data frame that defines the
the bivariate column. Defaults to "sex"}

\item{stack_by}{the name of the column in the data frame to use for shading
the bars}

\item{vertical_lines}{If you would like to add dashed vertical lines to help
visual interpretation of numbers. Default is to not show (FALSE),
to turn on write TRUE.}
}
\description{
Plot a population pyramid (age-sex) from a dataframe.
}
\examples{
library(ggplot2)

set.seed(2018-01-15)
ages <- cut(sample(80, 150, replace = TRUE),
            breaks = c(0, 5, 10, 30, 90), right = FALSE)
sex  <- sample(c("Female", "Male"), 150, replace = TRUE)
ill  <- sample(c("case", "non-case"), 150, replace = TRUE)
dat  <- data.frame(AGE = ages, sex = sex, ill = ill, stringsAsFactors = FALSE)

# Create the age pyramid, stratifying by sex
print(ap   <- plot_age_pyramid(dat, age_group = "AGE"))

# Stratify by case definition and customize with ggplot2
ap   <- plot_age_pyramid(dat, age_group = "AGE", split_by = "ill") +
  theme_bw(base_size = 16) +
  labs(title = "Age groups by case definition")
print(ap)

# Stratify by multiple factors
ap   <- plot_age_pyramid(dat, 
                         age_group = "AGE", 
                         split_by = "sex", 
                         stack_by = "ill",
                         vertical_lines = TRUE) +
  labs(title = "Age groups by case definition and sex")
print(ap)
}
