% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.R
\name{sample_size}
\alias{sample_size}
\title{Compute a sample size for surveys}
\usage{
sample_size(population_size, expected_prevalence, precision, design_effect,
  alpha = 0.05)
}
\description{
Compute a sample size for surveys
}
