% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-size.R
\name{sample_size_households}
\alias{sample_size_households}
\title{Take a sample size if children and convert it to households}
\usage{
sample_size_households(sample_size, avg_hh, prop_under_5, frac_6_59,
  non_response_rate = 0)
}
\description{
Take a sample size if children and convert it to households
}
