% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep-package.R
\docType{package}
\name{sitrep-package}
\alias{sitrep}
\alias{sitrep-package}
\title{sitrep: Report templates and helper functions for applied epidemiology}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Report templates and helper functions for applied epidemiology
}
\seealso{
Useful links:
\itemize{
  \item \url{https://r4epis.netlify.com}
  \item \url{https://github.com/R4EPI/sitrep}
  \item \url{https://r4epi.github.io/sitrep/}
  \item Report bugs at \url{https://github.com/R4EPI/sitrep/issues}
}

}
\author{
\strong{Maintainer}: Zhian N. Kamvar \email{zkamvar@gmail.com} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID})

Authors:
\itemize{
  \item Dirk Schumacher \email{mai@dirk-schumacher.net}
  \item Alexander Spina \email{alexfspina@gmail.com} (\href{https://orcid.org/0000-0001-8425-1867}{ORCID})
  \item Kate Doyle
  \item Chris Jarvis
  \item Lukas Richter
}

Other contributors:
\itemize{
  \item Annick Lenglet [copyright holder]
  \item Neale Batra [contributor]
}

}
\keyword{internal}
