% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sitrep-package.R
\docType{package}
\name{sitrep-package}
\alias{sitrep}
\alias{sitrep-package}
\title{sitrep: Report Templates and Helper Functions for Applied Epidemiology}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A meta-package that loads the complete sitrep ecosystem for applied epidemiology analysis. This package provides report templates and automatically loads companion packages, including 'epitabulate' (for epidemiological tables), 'epidict' (for data dictionaries), 'epikit' (for epidemiological utilities), and 'apyramid' (for age-sex pyramids). Simply load 'sitrep' to access all functions from the ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/R4EPI/sitrep/}
  \item \url{https://r4epi.github.io/sitrep/}
  \item Report bugs at \url{https://github.com/R4EPI/sitrep/issues/}
}

}
\author{
\strong{Maintainer}: Alexander Spina \email{aspina@appliedepi.org} (\href{https://orcid.org/0000-0001-8425-1867}{ORCID})

Authors:
\itemize{
  \item Zhian N. Kamvar \email{zkamvar@gmail.com} (\href{https://orcid.org/0000-0003-1458-7108}{ORCID})
  \item Dirk Schumacher \email{mai@dirk-schumacher.net}
  \item Kate Doyle
  \item Chris Jarvis
  \item Lukas Richter
  \item Paula Blomquist
}

Other contributors:
\itemize{
  \item Annick Lenglet [contributor]
  \item Neale Batra [contributor]
  \item Applied Epi Incorporated [copyright holder]
  \item Medecins Sans Frontieres Operational Centre Amsterdam [funder]
}

}
\keyword{internal}
