% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitabulate_exports.R
\docType{data}
\name{tab_survey}
\alias{tab_survey}
\alias{tab_linelist}
\alias{tab_univariate}
\alias{data_frame_from_2x2}
\title{Functions re-expored from {epitabulate}}
\format{
An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.

An object of class \code{function} of length 1.
}
\usage{
tab_survey

tab_linelist

tab_univariate

data_frame_from_2x2
}
\description{
Functions re-expored from {epitabulate}
}
\seealso{
\pkg{epitabulate} functions:
\itemize{
\item \code{\link[epitabulate:tab_functions]{epitabulate::tab_linelist()}}: tabulate linelist data
\item \code{\link[epitabulate:tab_functions]{epitabulate::tab_survey()}}: tabulate survey data
\item \code{\link[epitabulate:tab_univariate]{epitabulate::tab_univariate()}}: caluclate odds, risk, and incidence risk
ratios for multiple variables from linelist data.
\item \code{\link[epitabulate:data_frame_from_2x2]{epitabulate::data_frame_from_2x2()}}: convert a 2x2(x2) table to a data
frame clearly labelling the (un)exposed (non)case combinations and their
totals.
}
}
\keyword{datasets}
