% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epitabulate_exports.R
\docType{import}
\name{tab_survey}
\alias{tab_survey}
\alias{tab_linelist}
\alias{tab_univariate}
\alias{data_frame_from_2x2}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{epitabulate}{\code{\link[epitabulate]{data_frame_from_2x2}}, \code{\link[epitabulate:tab_functions]{tab_linelist}}, \code{\link[epitabulate:tab_functions]{tab_survey}}, \code{\link[epitabulate]{tab_univariate}}}
}}

