% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_univariate.R
\name{tab_univariate}
\alias{tab_univariate}
\title{Produce odds ratios, risk ratios or incidence rate ratios with accompanying confidence intervals}
\usage{
tab_univariate(x, outcome, exposure, perstime = NULL, strata = NULL,
  measure = "OR", extend_output = TRUE, digits = 3,
  mergeCI = FALSE, woolf_test = FALSE)
}
\arguments{
\item{x}{A data frame}

\item{outcome}{Name of A TRUE/FALSE variable as your outcome of interest (e.g. illness)}

\item{perstime}{A numeric variable containing the observation time for each individual}

\item{strata}{Name of a TRUE/FALSE variable to be used for stratifying results. Note that this results
in a different output table - giving you a table of crude measure, measures for each strata and
the mantel-haeszel adjusted measure for each exposure variable listed in \code{...}}

\item{measure}{Specify what you would like to calculated, options are "OR", "RR" or "IRR"
default is "OR"}

\item{extend_output}{TRUE/FALSE to specify whether would like all columns in the outputs (default is TRUE)
Non-extended output drops group odds or risk calculations as well as p-values}

\item{digits}{Specify number of decimal places (default is 3)}

\item{mergeCI}{Whether or not to put the confidence intervals in one column (default is FALSE)}

\item{woolf_test}{Only if strata specified and measure is "RR" or "OR". TRUE/FALSE to specify whether to
include woolf test for homogeneity p-value. Tests whether there is a significant difference in the
estimates between strata.}

\item{...}{Names of TRUE/FALSE variables as exposures of interest (e.g. risk factors)}
}
\description{
Produce odds ratios, risk ratios or incidence rate ratios with accompanying confidence intervals
}
\examples{
# generate a fake dataset
a <- tibble(potato = sample(c(TRUE, FALSE), 2000, replace = TRUE),
           apple = sample(c(TRUE, FALSE), 2000, replace = TRUE),
           orange = sample(c(TRUE, FALSE), 2000, replace = TRUE),
           stratifier = sample(c(TRUE, FALSE), 2000, replace = TRUE),
           perstime = sample(150:250, 2000, replace = TRUE)
           )

# get counts table
counts <- table(a[c("apple", "potato")])

## For odds ratios

# get odds of exposure among cases
expo_cases_odds <- counts[4]/counts[3]

# get odds of exposure among controls
expo_controls_odds <- counts[2]/counts[1]

# calculate odds ratio by using odds above
or_from_odds <- expo_cases_odds / expo_controls_odds

# calculate odds ratio by cross multiplying counts
or_from_cross <- (counts[4] * counts[1]) / (counts[2] * counts[3])

# get the results from tab_univariate function
func_res <- tab_univariate(a, potato, apple)

# tests
stopifnot(
# check that function OR is equal to OR from odds
func_res$exp == or_from_odds,

# check that function OR is equal to OR from cross multiplying
func_res$exp == or_from_cross,

# check that function case_odds is equal to count odds
func_res$cases_odds == expo_cases_odds,
func_res$controls_odds == expo_controls_odds
)
}
\references{
Inspired by Daniel Gardiner,
see \href{https://github.com/DanielGardiner/UsefulFunctions/blob/efffde624d424d977651ed1a9ee4430cbf2b0d6f/single.variable.analysis.v0.3.R#L12}{github repo}
}
