% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tab_univariate.R
\name{tab_univariate}
\alias{tab_univariate}
\title{Produce odds ratios, risk ratios or incidence rate ratios}
\usage{
tab_univariate(x, outcome, ..., perstime = NULL, strata = NULL,
  measure = "OR", extend_output = TRUE, digits = 3,
  mergeCI = FALSE, woolf_test = FALSE)
}
\arguments{
\item{x}{A data frame}

\item{outcome}{Name of A TRUE/FALSE variable as your outcome of interest
(e.g. illness)}

\item{...}{Names of TRUE/FALSE variables as exposures of interest (e.g. risk
factors)}

\item{perstime}{A numeric variable containing the observation time for each
individual}

\item{strata}{Name of a TRUE/FALSE variable to be used for stratifying
results. Note that this results in a different output table - giving you a
table of crude measure, measures for each strata and the mantel-haeszel
adjusted measure for each exposure variable listed in \code{...}}

\item{measure}{Specify what you would like to calculated, options are "OR",
"RR" or "IRR" default is "OR"}

\item{extend_output}{TRUE/FALSE to specify whether would like all columns in
the outputs (default is TRUE) Non-extended output drops group odds or risk
calculations as well as p-values}

\item{digits}{Specify number of decimal places (default is 3)}

\item{mergeCI}{Whether or not to put the confidence intervals in one column
(default is FALSE)}

\item{woolf_test}{Only if strata specified and measure is "RR" or "OR".
TRUE/FALSE to specify whether to include woolf test for homogeneity
p-value. Tests whether there is a significant difference in the estimates
between strata.}
}
\description{
Produce odds ratios, risk ratios or incidence rate ratios
}
\examples{

# generate a fake dataset
a <- data.frame(case_def = sample(c(TRUE, FALSE), 2000, replace = TRUE),
           riskA = sample(c(TRUE, FALSE), 2000, replace = TRUE),
           riskB = sample(c(TRUE, FALSE), 2000, replace = TRUE),
           stratifier = sample(c(TRUE, FALSE), 2000, replace = TRUE),
           perstime = sample(150:250, 2000, replace = TRUE)
           )

# get the results from tab_univariate function
func_res <- tab_univariate(a, case_def, riskA,
                           strata = stratifier, digits = 6, measure = "OR")

# get risk ratios
func_res <- tab_univariate(a, case_def, riskA,
                           strata = stratifier, digits = 6, measure = "RR")


}
\references{
Inspired by Daniel Gardiner,
see \href{https://github.com/DanielGardiner/UsefulFunctions/tree/master/single.variable.analysis.v0.3.R}{github repo}
}
