% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_survey.R
\name{tabulate_survey}
\alias{tabulate_survey}
\alias{tabulate_binary_survey}
\title{Tabulate survey design objects by a categorical and another stratifying variable}
\usage{
tabulate_survey(x, var, strata = NULL, pretty = TRUE, wide = TRUE,
  digits = 1, method = "logit", deff = FALSE, proptotal = FALSE,
  rowtotals = FALSE, coltotals = FALSE)

tabulate_binary_survey(x, ..., keep = NULL, invert = FALSE,
  pretty = TRUE, wide = TRUE, digits = 1, method = "logit",
  deff = FALSE)
}
\arguments{
\item{x}{a survey design object}

\item{var}{the bare name of a categorical variable}

\item{strata}{a variable to stratify the results by}

\item{pretty}{if \code{TRUE}, default, the proportion and CI are merged}

\item{wide}{if \code{TRUE} (default) and strata is defined, then the results are
presented in a wide table with each stratification counts and estimates in
separate columns. If \code{FALSE}, then the data will be presented in a long
format where the counts and estimates are presented in single columns. This
has no effect if strata is not defined.}

\item{digits}{if \code{pretty = FALSE}, this indicates the number of digits used
for proportion and CI}

\item{method}{a method from \code{\link[survey:svyciprop]{survey::svyciprop()}} to calculate the confidence
interval. Defaults to "logit"}

\item{deff}{a logical indicating if the design effect should be reported.
Defaults to "TRUE"}

\item{proptotal}{if \code{TRUE} and \code{strata} is not \code{NULL}, then the totals of the
rows will be reported as proportions of the total data set, otherwise, they
will be proportions within the stratum (default).}

\item{rowtotals}{if \code{TRUE} and \code{strata} is defined, then an extra "Total"
column will be added tabulating all of the rows across strata.}

\item{coltotals}{if \code{TRUE} a new row with totals for each "n" column is
created.}

\item{...}{binary variables for tabulation}

\item{keep}{a vector of binary values to keep}

\item{invert}{if \code{TRUE}, the kept values are rejected. Defaults to \code{FALSE}}
}
\value{
a long or wide tibble with tabulations n, ci, and deff
}
\description{
Tabulate survey design objects by a categorical and another stratifying variable
}
\note{
The proportions presented here represent the proportions of the total
population, not for the stratified samples
}
\examples{
library(srvyr)
library(survey)
data(api)

# stratified sample
s <- apistrat \%>\%
  as_survey_design(strata = stype, weights = pw) \%>\%
  tabulate_survey(stype, awards, coltotals = TRUE, rowtotals = TRUE, deff = TRUE)
s

# making things pretty
s \%>\%
  # wrap all "n" variables in braces (note space before n).
  augment_redundant(" n" = " (n)") \%>\%
  # relabel all columns containing "prop" to "\% (95\% CI)"
  rename_redundant("ci"   = "\% (95\% CI)",
                   "deff" = "Design Effect")

# long data
apistrat \%>\%
  as_survey_design(strata = stype, weights = pw) \%>\%
  tabulate_survey(stype, awards, wide = FALSE)

apistrat \%>\%
  as_survey_design(strata = stype, weights = pw) \%>\%
  tabulate_binary_survey(stype, awards, keep = c("Yes", "E"))

apistrat \%>\%
  as_survey_design(strata = stype, weights = pw) \%>\%
  tabulate_binary_survey(stype, awards, keep = c("Yes", "E"), invert = TRUE)
}
\seealso{
\code{\link[=rename_redundant]{rename_redundant()}}, \code{\link[=augment_redundant]{augment_redundant()}}
}
