% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabulate_death.R
\name{tabulate_survey}
\alias{tabulate_survey}
\title{Tabulate survey design objects by a categorical and another stratifying variable}
\usage{
tabulate_survey(x, var, strata = NULL, pretty = TRUE, digits = 1)
}
\arguments{
\item{x}{a survey design object}

\item{var}{the bare name of a categorical variable}

\item{strata}{a variable to stratify the results by}

\item{pretty}{if \code{TRUE}, default, the proportion and CI are merged}

\item{digits}{if \code{prittey = FALSE}, this indicates the number of digits used
for proportion and CI}
}
\value{
a tibble
}
\description{
Tabulate survey design objects by a categorical and another stratifying variable
}
\examples{
library(srvyr)
library(survey)
data(api)

# stratified sample
apistrat \%>\%
  as_survey_design(strata = stype, weights = pw) \%>\%
  tabulate_survey(stype, awards)

}
