% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariate-analysis.R
\name{univariate_analysis}
\alias{univariate_analysis}
\title{Produce risk ratios or odds ratios}
\usage{
univariate_analysis(outcome, ...)
}
\description{
Produce risk ratios or odds ratios
}
\references{
Inspired by Daniel Gardiner, 
see [github repo](https://github.com/DanielGardiner/UsefulFunctions/blob/efffde624d424d977651ed1a9ee4430cbf2b0d6f/single.variable.analysis.v0.3.R#L12)
}
