% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zcurve.R
\name{zcurve}
\alias{zcurve}
\title{Create a curve comparing observed Z-scores to the WHO standard.}
\usage{
zcurve(zscore)
}
\arguments{
\item{zscore}{a numeric vector containing computed zscores}
}
\value{
a ggplot2 object
}
\description{
Create a curve comparing observed Z-scores to the WHO standard.
}
\examples{
dat <- rnorm(204) + runif(1) # slightly skewed
zcurve(dat)
}
